分离爬虫与数据库接口
